reorganize database package into backend api
