+++ src/API/index.js
@@ -1,0 +1,8 @@
+export { db, auth, EmailAuthProvider } from './Firebase';
+export {
+  addParticipant,
+  addIncident,
+  addEngagement,
+  getParticipantById,
+  getIncidentsByParticipantId,
+} from './FirebaseModel';
