--- src/Database/index.js
@@ -1,4 +1,0 @@
-export { db, auth, EmailAuthProvider } from './Firebase';
-export { addEngagement, getParticipantById, getIncidentsByParticipantId } from './FirebaseEngagementForm'
-export { addIncident } from './FirebaseIncidentForm'
-export { addParticipant } from './FirebaseParticipantForm'
