+++ src/RecordProfileForm.js
@@ -1,7 +1,7 @@
 import React, { useReducer, useEffect, useMemo } from 'react';
 import Form from 'react-bootstrap/Form';
 import Button from 'react-bootstrap/Button';
-import { db } from './Firebase';
+import { db } from './API';
 import { useAuthState } from './Auth';
 
 export default () => {
